Protected Route and user context
Added protected routes depending from the user connection, implemented connection context so that login and logout are implemented. Adapted some style from pages which display different component if user is connected or not. For the moment, change the city location and climbing location permission to AllowAny in order to access the map even if a user isn't connected